Rename security_sdk to trifle for jvm / android. Move protos to root
