--- android/sample_app/src/main/java/app/cash/security_sdk/MainActivity.kt
@@ -1,22 +1,0 @@
-package app.cash.security_sdk
-
-import android.os.Bundle
-import android.util.Log
-import androidx.appcompat.app.AppCompatActivity
-import app.cash.security_sdk.sample_app.R
-
-class MainActivity : AppCompatActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-
-    val version = LibraryVersion()
-    Log.v(TAG, "Security SDK version: $version")
-
-    setContentView(R.layout.activity_main)
-  }
-
-  private companion object {
-    private const val TAG = "MainActivity"
-  }
-} 
-
