+++ android/sample_app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="app.cash.security_sdk.MainActivity">
+    tools:context="app.cash.trifle.MainActivity">
   <TextView
       android:layout_width="wrap_content"
       android:layout_height="wrap_content"
