+++ android/trifle/proguard-rules.pro
