--- jvm/src/main/kotlin/app/cash/s2dk/common/Version.kt
@@ -1,9 +1,0 @@
-package app.cash.s2dk.common
-
-interface Version {
-  fun complete(): String
-
-  fun major(): Int
-
-  fun minor(): Int
-}
