--- jvm/src/main/kotlin/app/cash/security_sdk/SigningCert.kt
@@ -1,5 +1,0 @@
-package app.cash.security_sdk
-
-import okio.ByteString
-
-data class SigningCert(val certificate: ByteString)
