--- jvm/src/main/kotlin/app/cash/security_sdk/internal/providers/TrifleContentVerifierProvider.kt
@@ -1,5 +1,0 @@
-package app.cash.security_sdk.internal.providers
-
-import org.bouncycastle.operator.ContentVerifierProvider
-
-sealed class TrifleContentVerifierProvider : ContentVerifierProvider
