+++ jvm/src/main/kotlin/app/cash/trifle/SigningCert.kt
@@ -1,0 +1,5 @@
+package app.cash.trifle
+
+import okio.ByteString
+
+data class SigningCert(val certificate: ByteString)
