+++ jvm/src/main/kotlin/app/cash/trifle/common/Version.kt
@@ -1,0 +1,9 @@
+package app.cash.trifle.common
+
+interface Version {
+  fun complete(): String
+
+  fun major(): Int
+
+  fun minor(): Int
+}
