+++ jvm/src/main/kotlin/app/cash/trifle/internal/providers/TrifleContentVerifierProvider.kt
@@ -1,0 +1,5 @@
+package app.cash.trifle.internal.providers
+
+import org.bouncycastle.operator.ContentVerifierProvider
+
+sealed class TrifleContentVerifierProvider : ContentVerifierProvider
