+++ jvm/src/main/kotlin/app/cash/trifle/internal/signers/TrifleContentSigner.kt
@@ -1,0 +1,8 @@
+package app.cash.trifle.internal.signers
+
+import org.bouncycastle.asn1.x509.SubjectPublicKeyInfo
+import org.bouncycastle.operator.ContentSigner
+
+sealed class TrifleContentSigner : ContentSigner {
+  abstract fun subjectPublicKeyInfo(): SubjectPublicKeyInfo
+}
