--- jvm/src/main/proto/squareup/cash/s2dk/api/alpha/public.proto
@@ -1,42 +1,0 @@
-syntax = "proto2";
-package squareup.cash.s2dk.api.alpha;
-
-option java_package = "com.squareup.protos.cash.s2dk.api.alpha";
-
-/**
- * The certificate request object generated by s2dk on mobile clients. This proto is used as a
- * serialization/deserialization mechanism for an otherwise opaque object whose representation is
- * internal to the library only. This primary purpose of this message is to convey the public key
- * portion of the key pair created on a client's device so that an authorization service may then
- * grant the client a certificate.
- */
-message MobileCertificateRequest {
-  // Version describing the current format of the MobileCertificateRequest.
-  // Required.
-  optional uint32 version = 1;
-
-  // Bytes representing a Certificate Request as specified in the PKCS10 RFC, see
-  // https://datatracker.ietf.org/doc/html/rfc5967 for details.
-  optional bytes pkcs10_request = 2;
-}
-
-message MobileCertificateResponse {
-  repeated bytes certificates = 1;
-}
-
-/**
- * The certificate object handled by s2dk. This proto is used as a serialization/deserialization
- * mechanism for an otherwise opaque object whose representation is internal to the library only.
- * The purpose of this message is to bind a public key with a set of verified attributes about the
- * entity which controls the private key corresponding to the given public key.
- */
-message Certificate {
-  // Version describing the current format of the underlying bytes of the certificate. This tells
-  // the s2dk library how to interpret the certificate bytes.
-  // Required.
-  optional uint32 version = 1;
-
-  // The current representation is fundamentally an x.509 certificate as defined in
-  // https://datatracker.ietf.org/doc/html/rfc5280, with most of the fields and features ignored.
-  optional bytes certificate = 2;
-}
