+++ jvm/src/test/kotlin/app/cash/trifle/common/LibraryVersionTest.kt
@@ -1,0 +1,30 @@
+package app.cash.trifle.common
+
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+internal class LibraryVersionUnitTest {
+  @Test
+  fun testCurrentVersionCanBeExtracted() {
+    val libraryVersion = LibraryVersion()
+    val versionString = libraryVersion.complete()
+
+    assertEquals(versionString, "0.1.2")
+  }
+
+  @Test
+  fun testMajorVersionExtracts() {
+    val libraryVersion = LibraryVersion()
+    val majorVersion = libraryVersion.major()
+
+    assertEquals(majorVersion, 0)
+  }
+
+  @Test
+  fun testMinorVersionExtracts() {
+    val libraryVersion = LibraryVersion()
+    val minorVersion = libraryVersion.minor()
+
+    assertEquals(minorVersion, 1)
+  }
+}
